Refactor: Use modules to link the code of each JS file
