+++ src/components/todo-list/toDoItems/toDoHandlers/editToDo/editToDo.css
@@ -1,0 +1,3 @@
+.itemInput:focus {
+    outline-color:rgb(7, 138, 7);
+}
