+++ todolist.html
@@ -7,8 +7,7 @@
     <link rel="stylesheet" href="src/reset.css" />
     <link rel="stylesheet" href="src/components/todo-list/toDoList.css" />
     <link rel="stylesheet" href="https://fonts.googleapis.com/css2?family=Material+Symbols+Outlined:opsz,wght,FILL,GRAD@20..48,100..700,0..1,-50..200" />
-    <script  type="module" src="src/components/todo-list/toDoList.js"></script>
-    <script  type="module" src="src/components/todo-list/mainInput/mainInput.js"></script>
+    <script type="module" src="src/components/todo-list/toDoList.js"></script>
     <title>ToDoList</title>
   </head>
   <body>
